Migrate baremetal inventory to a common location
With the introduction of the Ironic inventory API, these types are now
shared between Ironic and Inspector.
